4.21*: bloglist expansion, step9

+++ controllers/blogs.js
@@ -17,7 +17,7 @@ blogsRouter.post('/', async (request, response) => {
       response.status(400).end()
       return
     }
-    console.log(request.token)
+    
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
     if(!request.token || !decodedToken.id) {
       return response.status(401).json({error: 'Token missing or invalid'})
@@ -49,8 +49,20 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const blog = await Blog.findById(request.params.id)
+  
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if(!request.token || !decodedToken.id) {
+    return response.status(401).json({error: 'Token missing or invalid'})
+  }
+  const user = await User.findById(decodedToken.id)
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }
+  else {
+    return response.status(401).json({error: 'A blog can only be deleted by its creator'})
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
